chore: refactor code & fix tests
