+++ src/authentication/DeviceDescription.ts
@@ -1,0 +1,7 @@
+export type DeviceDescription =
+  'browser-chrome' |
+  'desktop-macos' |
+  'desktop-windows' |
+  'mobile-android' |
+  'mobile-ios' |
+  'remarkable'
