+++ test/records/Device .connect given valid pair token, updates Device with new session token/recordings_613089741/recording.har
@@ -40,7 +40,7 @@
           "content": {
             "mimeType": "text/plain; charset=utf-8",
             "size": 957,
-            "text": "eyJhbGciOiJIUzI1NiIsImtpZCI6IjEiLCJ0eXAiOiJKV1QifQ.eyJhdXRoMC1wcm9maWxlIjp7IlVzZXJJRCI6ImF1dGgwfDVlZmMyODdmYzQ4ODAxMDAxMzQ0ZjkyMyIsIklzU29jaWFsIjpmYWxzZSwiQ29ubmVjdGlvbiI6IlVzZXJuYW1lLVBhc3N3b3JkLUF1dGhlbnRpY2F0aW9uIiwiTmFtZSI6InBhc2N1MjE2QGdtYWlsLmNvbSIsIk5pY2tuYW1lIjoicGFzY3UyMTYiLCJHaXZlbk5hbWUiOiIiLCJGYW1pbHlOYW1lIjoiIiwiRW1haWwiOiJwYXNjdTIxNkBnbWFpbC5jb20iLCJFbWFpbFZlcmlmaWVkIjp0cnVlLCJDcmVhdGVkQXQiOiIyMDIwLTA3LTAxVDA2OjA5OjAzLjk0M1oiLCJVcGRhdGVkQXQiOiIyMDI0LTAyLTI2VDA3OjMzOjEzLjI4NloifSwiYmV0YSI6dHJ1ZSwiZGV2aWNlLWRlc2MiOiJicm93c2VyLWNocm9tZSIsImRldmljZS1pZCI6IjAyY2U3OTUwLTBiMWUtNDAzOS05NWE3LWUwOThlMTBjMzNmYSIsImV4cCI6MTcwOTAyNzc3NiwiaWF0IjoxNzA5MDE2OTc2LCJpc3MiOiJyTSBXZWJBcHAiLCJqdGkiOiJjazB0QU41Ym1OMD0iLCJsZXZlbCI6ImNvbm5lY3QiLCJuYmYiOjE3MDkwMTY5NzYsInNjb3BlcyI6InN5bmM6dG9ydG9pc2Ugc2NyZWVuc2hhcmUgaW50Z3IgaHdjbWFpbDotMSBtYWlsOi0xIGRvY2VkaXQiLCJzdWIiOiJhdXRoMHw1ZWZjMjg3ZmM0ODgwMTAwMTM0NGY5MjMifQ.XaXDgLv_tskCyyorOJUaiHdEDRkcOk8BTm6jja_rDlk"
+            "text": "eyJhbGciOiJIUzI1NiIsImtpZCI6IjEiLCJ0eXAiOiJKV1QifQ.eyJhdXRoMC1wcm9maWxlIjp7IlVzZXJJRCI6ImF1dGgwfDVlZmMyODdmYzQ4ODAxMDAxMzQ0ZjkyMyIsIklzU29jaWFsIjpmYWxzZSwiQ29ubmVjdGlvbiI6IlVzZXJuYW1lLVBhc3N3b3JkLUF1dGhlbnRpY2F0aW9uIiwiTmFtZSI6InBhc2N1MjE2QGdtYWlsLmNvbSIsIk5pY2tuYW1lIjoicGFzY3UyMTYiLCJHaXZlbk5hbWUiOiIiLCJGYW1pbHlOYW1lIjoiIiwiRW1haWwiOiJwYXNjdTIxNkBnbWFpbC5jb20iLCJFbWFpbFZlcmlmaWVkIjp0cnVlLCJDcmVhdGVkQXQiOiIyMDIwLTA3LTAxVDA2OjA5OjAzLjk0M1oiLCJVcGRhdGVkQXQiOiIyMDI0LTAyLTI2VDA3OjMzOjEzLjI4NloifSwiYmV0YSI6dHJ1ZSwiZGV2aWNlLWRlc2MiOiJicm93c2VyLWNocm9tZSIsImRldmljZS1pZCI6IjAyY2U3OTUwLTBiMWUtNDAzOS05NWE3LWUwOThlMTBjMzNmYSIsImV4cCI6MTcwOTAyODY2NiwiaWF0IjoxNzA5MDE3ODY2LCJpc3MiOiJyTSBXZWJBcHAiLCJqdGkiOiJjazB0K29CWnhXTT0iLCJsZXZlbCI6ImNvbm5lY3QiLCJuYmYiOjE3MDkwMTc4NjYsInNjb3BlcyI6InN5bmM6dG9ydG9pc2Ugc2NyZWVuc2hhcmUgZG9jZWRpdCBpbnRnciBtYWlsOi0xIGh3Y21haWw6LTEiLCJzdWIiOiJhdXRoMHw1ZWZjMjg3ZmM0ODgwMTAwMTM0NGY5MjMifQ.DLB0lT7ONr8Du3JZrKPjHsyJ1GrjSb2gVL4Ig6LJ624"
           },
           "cookies": [],
           "headers": [
@@ -54,11 +54,11 @@
             },
             {
               "name": "x-ratelimit-remaining",
-              "value": "3"
+              "value": "4"
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "13"
+              "value": "12"
             },
             {
               "name": "content-type",
@@ -66,11 +66,11 @@
             },
             {
               "name": "x-cloud-trace-context",
-              "value": "81eb05aedb7102392a4601919c890994"
+              "value": "f17a46ed6b1a81eda923ea7676bd8bd0"
             },
             {
               "name": "date",
-              "value": "Tue, 27 Feb 2024 06:56:16 GMT"
+              "value": "Tue, 27 Feb 2024 07:11:06 GMT"
             },
             {
               "name": "server",
@@ -95,8 +95,8 @@
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-27T06:56:13.186Z",
-        "time": 3088,
+        "startedDateTime": "2024-02-27T07:11:02.914Z",
+        "time": 4063,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -104,7 +104,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 3088
+          "wait": 4063
         }
       }
     ],
