+++ test/records/FetchClient .delete performs DELETE request/recordings_613089741/recording.har
@@ -52,7 +52,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:28 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:44 GMT"
             },
             {
               "name": "content-length",
@@ -64,11 +64,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943548&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=KSllafMwFIY6gS1i1DP%2FfrmCsGVCHrS%2B4blV5NthVig%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017484&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=LRl9T7vO90FvDUiNDxiYFSoIGe2f1GXQHXrDXXR6o8Y%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943548&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=KSllafMwFIY6gS1i1DP%2FfrmCsGVCHrS%2B4blV5NthVig%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017484&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=LRl9T7vO90FvDUiNDxiYFSoIGe2f1GXQHXrDXXR6o8Y%3D"
             },
             {
               "name": "nel",
@@ -88,7 +88,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943579"
+              "value": "1709017488"
             },
             {
               "name": "access-control-allow-origin",
@@ -120,21 +120,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b7863abddd86c8-MAD"
+              "value": "85be934fa86d2fa6-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1015,
+          "headersSize": 1007,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 204,
           "statusText": "No Content"
         },
-        "startedDateTime": "2024-02-26T10:32:28.507Z",
-        "time": 399,
+        "startedDateTime": "2024-02-27T07:04:42.677Z",
+        "time": 2138,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -142,7 +142,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 399
+          "wait": 2138
         }
       },
       {
@@ -199,7 +199,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:29 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:45 GMT"
             },
             {
               "name": "content-type",
@@ -215,11 +215,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943549&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=JlnMXMFZ745tz3PyrKSSE%2Fv3ug6YoC44TuS%2FaUrqTxY%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017485&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=pLCUdP902SLN1zbY5ZvnxV2ula1Fp87Ve2l7Nf56NTw%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943549&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=JlnMXMFZ745tz3PyrKSSE%2Fv3ug6YoC44TuS%2FaUrqTxY%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017485&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=pLCUdP902SLN1zbY5ZvnxV2ula1Fp87Ve2l7Nf56NTw%3D"
             },
             {
               "name": "nel",
@@ -239,7 +239,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943579"
+              "value": "1709017488"
             },
             {
               "name": "access-control-allow-origin",
@@ -287,21 +287,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b7863d09c086c8-MAD"
+              "value": "85be935069462fa6-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1115,
+          "headersSize": 1107,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-26T10:32:28.912Z",
-        "time": 139,
+        "startedDateTime": "2024-02-27T07:04:44.819Z",
+        "time": 290,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -309,7 +309,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 139
+          "wait": 290
         }
       }
     ],
