+++ test/records/FetchClient .get performs GET request/recordings_613089741/recording.har
@@ -62,7 +62,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:27 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:37 GMT"
             },
             {
               "name": "content-type",
@@ -146,7 +146,7 @@
             },
             {
               "name": "age",
-              "value": "5554"
+              "value": "3607"
             },
             {
               "name": "server",
@@ -154,7 +154,7 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b786315d472f93-MAD"
+              "value": "85be9323b8e56641-MAD"
             },
             {
               "name": "content-encoding",
@@ -171,8 +171,8 @@
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-26T10:32:26.979Z",
-        "time": 87,
+        "startedDateTime": "2024-02-27T07:04:34.011Z",
+        "time": 3705,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -180,7 +180,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 87
+          "wait": 3705
         }
       }
     ],
