+++ test/records/FetchClient .patch performs PATCH request/recordings_613089741/recording.har
@@ -52,7 +52,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:27 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:42 GMT"
             },
             {
               "name": "content-length",
@@ -64,11 +64,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943547&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=W6lfb0E3L3pUBeVOIMv4CKPiqB68MDc2njuwsXwFEec%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017481&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=4gGMFx9RGtDQKb8MKGBB6pmACSPnCvHjbcv5DYTn%2Frc%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943547&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=W6lfb0E3L3pUBeVOIMv4CKPiqB68MDc2njuwsXwFEec%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017481&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=4gGMFx9RGtDQKb8MKGBB6pmACSPnCvHjbcv5DYTn%2Frc%3D"
             },
             {
               "name": "nel",
@@ -88,7 +88,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943579"
+              "value": "1709017488"
             },
             {
               "name": "access-control-allow-origin",
@@ -120,21 +120,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b786343b8f74de-MAD"
+              "value": "85be933d2a4e2f97-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1007,
+          "headersSize": 1011,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 204,
           "statusText": "No Content"
         },
-        "startedDateTime": "2024-02-26T10:32:27.450Z",
-        "time": 407,
+        "startedDateTime": "2024-02-27T07:04:40.977Z",
+        "time": 1044,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -142,7 +142,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 407
+          "wait": 1044
         }
       },
       {
@@ -207,13 +207,13 @@
             "encoding": "base64",
             "mimeType": "application/json; charset=utf-8",
             "size": 287,
-            "text": "[\"H4sIAAAAAAAAA1SPS2oEQQxD93MKUetssp0b5A69cVwKY+j6jMseJoTcPXRDCAEjJJAf6OsClFz0t1queH05ov3ZsNhZrigre0Ay8CFKJ5yT+y6B6eNhlT0wVIUqYeBTOSPdMGbYOMrOG3ulW5QT/D7q58G9pwkYWLnUpsXWfx2cmkt6FUJHX7xn9kgHn5PV4nzTbFv/h0cbO1eYEBnH3VMWYoS0rfexwrPBeSjXOYkN5zpnAx/sxoC43iyoMcrl+wcAAP//AwAVxpPoJAE=\",\"AAA=\"]"
+            "text": "[\"H4sIAAAAAAAAA1SPS2oEQQxD93MKUetssp0b5A69cVwKY+j6jMseJoTcPXRDCAEjJJAf6OsClFz0t1queH05ov3ZsNhZrigre0Ay8CFKJ5yT+y6B6eNhlT0wVIUqYeBTOSPdMGbYOMrOG3ulW5QT/D7q58G9pwkYWLnUpsXWfx2cmkt6FUJHX7xn9kgHn5PV4nzTbFv/h0cbO1eYEBnH3VMWYoS0rfexwrPBeSjXOYkN5zpnAx/sxoC43iyoMQ==\",\"yuX7BwAA//8DABXGk+gkAQAA\"]"
           },
           "cookies": [],
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:28 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:42 GMT"
             },
             {
               "name": "content-type",
@@ -229,11 +229,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943547&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=W6lfb0E3L3pUBeVOIMv4CKPiqB68MDc2njuwsXwFEec%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017482&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=CQovzpx%2FZ0iDAatLfa%2Bb7%2FQglthSVjq8yE9o%2B0V4YVo%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943547&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=W6lfb0E3L3pUBeVOIMv4CKPiqB68MDc2njuwsXwFEec%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017482&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=CQovzpx%2FZ0iDAatLfa%2Bb7%2FQglthSVjq8yE9o%2B0V4YVo%3D"
             },
             {
               "name": "nel",
@@ -253,7 +253,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943579"
+              "value": "1709017488"
             },
             {
               "name": "access-control-allow-origin",
@@ -301,7 +301,7 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b786367e8674de-MAD"
+              "value": "85be933eec462f97-MAD"
             },
             {
               "name": "content-encoding",
@@ -312,14 +312,14 @@
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1142,
+          "headersSize": 1158,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-26T10:32:27.864Z",
-        "time": 111,
+        "startedDateTime": "2024-02-27T07:04:42.025Z",
+        "time": 114,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -327,7 +327,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 111
+          "wait": 114
         }
       }
     ],
