+++ test/records/FetchClient .post performs POST request/recordings_613089741/recording.har
@@ -74,7 +74,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:27 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:40 GMT"
             },
             {
               "name": "content-type",
@@ -90,11 +90,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943547&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=W6lfb0E3L3pUBeVOIMv4CKPiqB68MDc2njuwsXwFEec%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017480&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=RAJY%2FDha%2BVCdPw6asAU64FW04m2IOskV8CK4n2mXyRQ%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943547&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=W6lfb0E3L3pUBeVOIMv4CKPiqB68MDc2njuwsXwFEec%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017480&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=RAJY%2FDha%2BVCdPw6asAU64FW04m2IOskV8CK4n2mXyRQ%3D"
             },
             {
               "name": "nel",
@@ -114,7 +114,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943579"
+              "value": "1709017488"
             },
             {
               "name": "access-control-allow-origin",
@@ -170,21 +170,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b78631f9ae8684-MAD"
+              "value": "85be933698915e14-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1231,
+          "headersSize": 1239,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "https://jsonplaceholder.typicode.com/posts/101",
           "status": 201,
           "statusText": "Created"
         },
-        "startedDateTime": "2024-02-26T10:32:27.082Z",
-        "time": 346,
+        "startedDateTime": "2024-02-27T07:04:37.743Z",
+        "time": 3218,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -192,7 +192,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 346
+          "wait": 3218
         }
       }
     ],
