+++ test/records/FetchClient .put performs PUT request/recordings_613089741/recording.har
@@ -52,7 +52,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:28 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:42 GMT"
             },
             {
               "name": "content-length",
@@ -64,11 +64,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943548&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=KSllafMwFIY6gS1i1DP%2FfrmCsGVCHrS%2B4blV5NthVig%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017482&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=CQovzpx%2FZ0iDAatLfa%2Bb7%2FQglthSVjq8yE9o%2B0V4YVo%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943548&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=KSllafMwFIY6gS1i1DP%2FfrmCsGVCHrS%2B4blV5NthVig%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017482&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=CQovzpx%2FZ0iDAatLfa%2Bb7%2FQglthSVjq8yE9o%2B0V4YVo%3D"
             },
             {
               "name": "nel",
@@ -88,7 +88,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943579"
+              "value": "1709017488"
             },
             {
               "name": "access-control-allow-origin",
@@ -120,21 +120,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b7863799256611-MAD"
+              "value": "85be934038332faf-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1015,
+          "headersSize": 1023,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 204,
           "statusText": "No Content"
         },
-        "startedDateTime": "2024-02-26T10:32:27.988Z",
-        "time": 177,
+        "startedDateTime": "2024-02-27T07:04:42.149Z",
+        "time": 198,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -142,7 +142,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 177
+          "wait": 198
         }
       },
       {
@@ -212,7 +212,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:32:28 GMT"
+              "value": "Tue, 27 Feb 2024 07:04:42 GMT"
             },
             {
               "name": "content-type",
@@ -228,11 +228,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943548&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=KSllafMwFIY6gS1i1DP%2FfrmCsGVCHrS%2B4blV5NthVig%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017482&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=CQovzpx%2FZ0iDAatLfa%2Bb7%2FQglthSVjq8yE9o%2B0V4YVo%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943548&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=KSllafMwFIY6gS1i1DP%2FfrmCsGVCHrS%2B4blV5NthVig%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017482&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=CQovzpx%2FZ0iDAatLfa%2Bb7%2FQglthSVjq8yE9o%2B0V4YVo%3D"
             },
             {
               "name": "nel",
@@ -252,7 +252,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943579"
+              "value": "1709017488"
             },
             {
               "name": "access-control-allow-origin",
@@ -300,20 +300,20 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b786387a656611-MAD"
+              "value": "85be9340f8ed2faf-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1116,
+          "headersSize": 1124,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-26T10:32:28.170Z",
+        "startedDateTime": "2024-02-27T07:04:42.351Z",
         "time": 315,
         "timings": {
           "blocked": -1,
