+++ test/records/NodeClient .delete performs DELETE request/recordings_613089741/recording.har
@@ -37,7 +37,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:33:17 GMT"
+              "value": "Tue, 27 Feb 2024 07:05:30 GMT"
             },
             {
               "name": "content-type",
@@ -53,11 +53,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943597&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=5NfpyOORoYVT2%2FEyPXu3WahIO4u5r7WsB0ztmcc3NHA%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017530&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=AsKZJ0%2Bb2Frdw%2BjB8bUGU97PObhP00F3i1R4cu786PI%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943597&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=5NfpyOORoYVT2%2FEyPXu3WahIO4u5r7WsB0ztmcc3NHA%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017530&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=AsKZJ0%2Bb2Frdw%2BjB8bUGU97PObhP00F3i1R4cu786PI%3D"
             },
             {
               "name": "nel",
@@ -77,7 +77,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943639"
+              "value": "1709017548"
             },
             {
               "name": "vary",
@@ -121,21 +121,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b78769e89286bc-MAD"
+              "value": "85be946dd8db2f92-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1064,
+          "headersSize": 1068,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-26T10:33:17.050Z",
-        "time": 113,
+        "startedDateTime": "2024-02-27T07:05:30.500Z",
+        "time": 102,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -143,7 +143,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 113
+          "wait": 102
         }
       }
     ],
