+++ test/records/NodeClient .get performs GET request/recordings_613089741/recording.har
@@ -37,7 +37,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:33:16 GMT"
+              "value": "Tue, 27 Feb 2024 07:05:29 GMT"
             },
             {
               "name": "content-type",
@@ -53,11 +53,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708476221&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=9e%2BoOw%2Fe6tako5b%2BcKIrbLPd%2BJZBxrDeFtQ914BeJJ4%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708477539&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=aS7QFMVgMg1ixy638HFkL8ggyNzlrdzC5CQUFVbWJk0%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708476221&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=9e%2BoOw%2Fe6tako5b%2BcKIrbLPd%2BJZBxrDeFtQ914BeJJ4%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708477539&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=aS7QFMVgMg1ixy638HFkL8ggyNzlrdzC5CQUFVbWJk0%3D"
             },
             {
               "name": "nel",
@@ -77,7 +77,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708476281"
+              "value": "1708477541"
             },
             {
               "name": "vary",
@@ -117,7 +117,7 @@
             },
             {
               "name": "age",
-              "value": "557"
+              "value": "12281"
             },
             {
               "name": "accept-ranges",
@@ -129,21 +129,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b787673b9186bc-MAD"
+              "value": "85be946a7ca52f92-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1111,
+          "headersSize": 1097,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-26T10:33:16.558Z",
-        "time": 127,
+        "startedDateTime": "2024-02-27T07:05:29.892Z",
+        "time": 85,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -151,7 +151,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 127
+          "wait": 85
         }
       }
     ],
