+++ test/records/NodeClient .patch performs PATCH request/recordings_613089741/recording.har
@@ -42,7 +42,7 @@
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 26 Feb 2024 10:33:16 GMT"
+              "value": "Tue, 27 Feb 2024 07:05:30 GMT"
             },
             {
               "name": "content-type",
@@ -58,11 +58,11 @@
             },
             {
               "name": "report-to",
-              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1708943596&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=eoMu4vsYaLKj8jNJk9GeuqGw%2Bm4jhjR2qNbU58eXbkM%3D\"}]}"
+              "value": "{\"group\":\"heroku-nel\",\"max_age\":3600,\"endpoints\":[{\"url\":\"https://nel.heroku.com/reports?ts=1709017530&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=AsKZJ0%2Bb2Frdw%2BjB8bUGU97PObhP00F3i1R4cu786PI%3D\"}]}"
             },
             {
               "name": "reporting-endpoints",
-              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1708943596&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=eoMu4vsYaLKj8jNJk9GeuqGw%2Bm4jhjR2qNbU58eXbkM%3D"
+              "value": "heroku-nel=https://nel.heroku.com/reports?ts=1709017530&sid=e11707d5-02a7-43ef-b45e-2cf4d2036f7d&s=AsKZJ0%2Bb2Frdw%2BjB8bUGU97PObhP00F3i1R4cu786PI%3D"
             },
             {
               "name": "nel",
@@ -82,7 +82,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1708943639"
+              "value": "1709017548"
             },
             {
               "name": "vary",
@@ -126,21 +126,21 @@
             },
             {
               "name": "cf-ray",
-              "value": "85b787686df386bc-MAD"
+              "value": "85be946b5dd92f92-MAD"
             },
             {
               "name": "alt-svc",
               "value": "h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1068,
+          "headersSize": 1072,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2024-02-26T10:33:16.811Z",
-        "time": 110,
+        "startedDateTime": "2024-02-27T07:05:30.101Z",
+        "time": 101,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -148,7 +148,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 110
+          "wait": 101
         }
       }
     ],
